@@ -2,7 +2,7 @@ name: add_language
 description: An i18n app example that adds a supported language
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
